[DESK-15] Backup service
Created the Backup service
